requirements.txt
asgiref==3.2.10
bbcode==1.1.0
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-import-export==2.3.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
psycopg2==2.8.5
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0
